Contrast the expandIcon with tertiary
