+++ plugins/data-management/src/PluginStoreWidget/components/PluginStoreWidget.tsx
@@ -32,7 +32,7 @@ const useStyles = makeStyles()(theme => ({
     margin: theme.spacing(1),
   },
   expandIcon: {
-    color: '#fff',
+    color: theme.palette.tertiary.contrastText,
   },
   adminBadge: {
     margin: '0.5em',
